Refactor edit step into its own component and add logic for saving images
